+++ src/Rest/GroupProcessor.php
@@ -18,13 +18,11 @@ class GroupProcessor extends AbstractDataProcessor implements LoggerAwareInterfa
 {
     use LoggerAwareTrait;
 
-    private GroupService $groupService;
-    private AuthorizationService $authorizationService;
-
-    public function __construct(GroupService $groupService, AuthorizationService $authorizationService)
+    public function __construct(
+        private readonly GroupService $groupService,
+        private readonly AuthorizationService $authorizationService)
     {
-        $this->groupService = $groupService;
-        $this->authorizationService = $authorizationService;
+        parent::__construct();
     }
 
     protected function isCurrentUserAuthorizedToAddItem($item, array $filters): bool